добавляет модульность фалов js
